+++ saleor/checkout/complete_checkout.py
@@ -197,6 +197,7 @@ def _create_line_for_order(
         translated_product_name=translated_product_name,
         translated_variant_name=translated_variant_name,
         product_sku=variant.sku,
+        product_variant_id=variant.get_global_id(),
         is_shipping_required=variant.is_shipping_required(),
         is_gift_card=variant.is_gift_card(),
         quantity=quantity,