+++ saleor/graphql/order/tests/test_fulfillment_return_products.py
@@ -831,6 +831,7 @@ def test_fulfillment_return_products_fulfillment_lines_and_order_lines(
         product_name=str(variant.product),
         variant_name=str(variant),
         product_sku=variant.sku,
+        product_variant_id=variant.get_global_id(),
         is_shipping_required=variant.is_shipping_required(),
         is_gift_card=variant.is_gift_card(),
         quantity=quantity,