+++ CHANGELOG.md
@@ -271,6 +271,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Use root level channel argument for filtering and sorting - #7374 by @IKarbowiak
   - drop `channel` field from filters and sorters
 - Drop top-level `checkoutLine` query from the schema with related resolver, use `checkout` query instead - #7623 by @dexon44
+- Make SKU an optional field on `ProductVariant` - #7633 by @rafalp
 
 ### Other